+++ src/storage.js
@@ -1,3 +1,4 @@
-export function updateLocalStorage() {
-    localStorage.setItem('tasks', JSON.stringify(tasks));
-}
+module.exports = function updateLocalStorage() {
+  const tasks = JSON.parse(localStorage.getItem('tasks'));
+  localStorage.setItem('tasks', JSON.stringify(tasks));
+};
